Fix a bug when dealing with null default value for Array in a schema
A topic containing Avro records that were serialized using a schema containing
arrays of Strings whose default value was null couldn't be dumped anymore,
although the schema and the records were perfectly valid. NPE was thrown.

The issue was that when a field is expected to be an array, Connect tries to
iterate over that array, although the array might have a null default value
(which is perfectly valid). This caused a NPE to be thrown.

Simply checking the array against null fixes the issue.